JavaScript errors in wiki. #4189
The problem seems to be with wikis with annotations off. The JavaScript
assumes that the annotation icons are on the page, so causes a fatal
error when it tries to add event handlers to them. This prevents the
rest of the JS on the page from being correctly initialised.

Fix tweaked by Tim Hunt before being added to ouvle.