Possible fix for seperate group error
Create a wiki with the settings 'seperate groups' and 'seperate wiki for
each group'.
Open the wiki, you will get the message that the page has not been created
yet.
Reload the page and you will receive a db write error.

table: ouwikis_subwikis
I think the problem is related to two uniqueness restrictions:
1. There must be only one entry for a group and wiki
2. There must be only one entry for a wiki and user (a seperate wiki for
every user)

In case of group wikis the user field is NULL. When opening the wiki
first, an table entry is generated. When opening it the second time,
Moodle tries to generate a second entry. But there already is the same
combination of wiki and user (NULL)

- The user-id is written to every entry in the mentioned table to avoid
  collisions of wikiid<->NULL
- When getting the subwiki, moodle shall not look for user=NULL

This may lead to problems in case of seperate wikis for every user