Use moodle_url to generate links in wiki history page
In M3.4+, MDL-59388 has changed how the link_arrow_[left|right]
functions render URLs. If it's passed a plain url with pre-escaped
ampersands in it, these now get double escaped.

To remedy this, proper moodle_url objects should be used for handling
the printing of urls. This change updates the links used in the
wikihistory page to use that API for all its urls.

This way the code change is compatible with earlier versions of Moodle
as well as 3.4+.

There is no functional change for the end user.